Keep permission command parsing behind the shared renderer boundary
