+++ src/routes/Chat/permission-request.ts
@@ -11,6 +11,7 @@ import {
   permissionRequestHasSensitiveResource,
   permissionRequestNeedsDefaultPrompt,
   permissionCommand,
+  permissionDeletionTargets,
   permissionPrimaryResource,
   permissionRequestKind,
   requestMatchesSessionGrant,
@@ -27,6 +28,7 @@ export {
   permissionRequestHasSensitiveResource,
   permissionRequestNeedsDefaultPrompt,
   permissionCommand,
+  permissionDeletionTargets,
   permissionPrimaryResource,
   permissionRequestKind,
   requestMatchesSessionGrant,
